refactor: MemberRepository 패키지 구조 재구성 및 인증 로직을 member 패키지로 이동
- `MemberRepository`를 새로운 패키지 구조로 옮겼습니다.
- 인증 및 회원 관련 로직을 member 패키지로 통합하여 응집도를 높였습니다.
- 닉네임 변경 검증 로직과 로깅 포맷을 개선했습니다.
