+++ src/main/java/com/ureca/snac/asset/service/AssetHistoryServiceImpl.java
@@ -8,7 +8,7 @@ import com.ureca.snac.asset.event.AssetChangedEvent;
 import com.ureca.snac.asset.repository.AssetHistoryRepository;
 import com.ureca.snac.common.CursorResult;
 import com.ureca.snac.member.Member;
-import com.ureca.snac.member.MemberRepository;
+import com.ureca.snac.member.repository.MemberRepository;
 import com.ureca.snac.member.exception.MemberNotFoundException;
 import com.ureca.snac.payment.entity.Payment;
 import com.ureca.snac.payment.repository.PaymentRepository;
