+++ src/main/java/com/ureca/snac/auth/repository/AuthRepository.java
@@ -9,8 +9,6 @@ import java.util.Optional;
 
 @Repository
 public interface AuthRepository extends JpaRepository<Member, Long> {
-    Boolean existsByEmail(String email);
-    Boolean existsByNickname(String nickname);
 
     Optional<Member> findByEmail(String email);
