--- src/main/java/com/ureca/snac/auth/service/JoinService.java
@@ -1,7 +1,0 @@
-package com.ureca.snac.auth.service;
-
-import com.ureca.snac.auth.dto.request.JoinRequest;
-
-public interface JoinService {
-    void joinProcess(JoinRequest joinRequest);
-}
