+++ src/main/java/com/ureca/snac/dev/support/DevDataSupport.java
@@ -5,7 +5,7 @@ import com.ureca.snac.board.entity.constants.CardCategory;
 import com.ureca.snac.board.entity.constants.Carrier;
 import com.ureca.snac.board.repository.CardRepository;
 import com.ureca.snac.member.Member;
-import com.ureca.snac.member.MemberRepository;
+import com.ureca.snac.member.repository.MemberRepository;
 import com.ureca.snac.member.exception.MemberNotFoundException;
 import com.ureca.snac.money.entity.MoneyRecharge;
 import com.ureca.snac.money.repository.MoneyRechargeRepository;
