+++ src/main/java/com/ureca/snac/finance/service/AccountServiceImpl.java
@@ -10,7 +10,7 @@ import com.ureca.snac.finance.repository.AccountRepository;
 import com.ureca.snac.finance.repository.BankRepository;
 import com.ureca.snac.finance.service.response.AccountResponse;
 import com.ureca.snac.member.Member;
-import com.ureca.snac.member.MemberRepository;
+import com.ureca.snac.member.repository.MemberRepository;
 import com.ureca.snac.member.exception.MemberNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
