+++ src/main/java/com/ureca/snac/member/service/JoinService.java
@@ -1,0 +1,7 @@
+package com.ureca.snac.member.service;
+
+import com.ureca.snac.member.dto.request.JoinRequest;
+
+public interface JoinService {
+    void joinProcess(JoinRequest joinRequest);
+}
