+++ src/main/java/com/ureca/snac/mypage/service/MyPageServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ureca.snac.mypage.service;
 
 import com.ureca.snac.member.Member;
-import com.ureca.snac.member.MemberRepository;
+import com.ureca.snac.member.repository.MemberRepository;
 import com.ureca.snac.member.exception.MemberNotFoundException;
 import com.ureca.snac.mypage.dto.MyPageResponse;
 import lombok.RequiredArgsConstructor;
