+++ src/main/java/com/ureca/snac/payment/service/PaymentServiceImpl.java
@@ -2,7 +2,7 @@ package com.ureca.snac.payment.service;
 
 import com.ureca.snac.infra.PaymentGatewayAdapter;
 import com.ureca.snac.member.Member;
-import com.ureca.snac.member.MemberRepository;
+import com.ureca.snac.member.repository.MemberRepository;
 import com.ureca.snac.member.exception.MemberNotFoundException;
 import com.ureca.snac.payment.dto.PaymentCancelResponse;
 import com.ureca.snac.payment.dto.PaymentFailureRequest;
