+++ src/main/java/com/ureca/snac/trade/service/TradeQueryServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ureca.snac.trade.service;
 
 import com.ureca.snac.member.Member;
-import com.ureca.snac.member.MemberRepository;
+import com.ureca.snac.member.repository.MemberRepository;
 import com.ureca.snac.member.exception.MemberNotFoundException;
 import com.ureca.snac.trade.controller.request.TradeQueryType;
 import com.ureca.snac.trade.dto.TradeDto;
